pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "Application of Self-Supervised Point Cloud Deep Learning for Forest Biomass Estimation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "torch==2.5.1",
  "torchvision==0.20.1",
  "lightning==2.5.0",
  "ocnn==2.2.6",
  "scikit-learn==1.6.1",
  "notebook==7.3",
  "matplotlib==3.10",
  "seaborn==0.13.2",
  "laspy[lazrs]==2.5.4",
  "ray[data,train,tune]==2.41",
  "pandas==2.2.3",
  "geopandas==1.0.1",
  "rasterio==1.4.3",
  "rioxarray==0.18.2",
  "ipywidgets==8.1.5",
  "wandb==0.19.6",
  "toml==0.10.2",
  "matplotlib-scalebar==0.9.0",
  "cupy-cuda12x==13.4.1",
  "geocube==0.7.1",
  "pillow==11.2.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu124", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true
